# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hitler Mod app store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface App {
  id: string;
  name: string;
  category: string;
  description: string;
  logoUrl: string;
  screenshots: string[];
  downloadUrl: string;
  size: string;
  version: string;
  developer: string;
  rating: number;
  downloads: number;
  featured: boolean;
  createdAt: string;
}

export interface CreateAppBody {
  name: string;
  category: string;
  description: string;
  logoUrl: string;
  screenshots: string[];
  downloadUrl: string;
  size: string;
  version: string;
  developer: string;
  featured?: boolean;
}

export interface Category {
  slug: string;
  name: string;
  appCount: number;
}

export interface DownloadResponse {
  url: string;
  downloads: number;
}

export interface StatsOverview {
  totalApps: number;
  totalDownloads: number;
  totalCategories: number;
  averageRating: number;
}

export type ListAppsParams = {
  search?: string;
  category?: string;
  sort?: ListAppsSort;
};

export type ListAppsSort = (typeof ListAppsSort)[keyof typeof ListAppsSort];

export const ListAppsSort = {
  latest: "latest",
  trending: "trending",
  rating: "rating",
} as const;