# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hitler Mod app store API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: apps
  - name: categories
  - name: stats
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /apps:
    get:
      operationId: listApps
      tags: [apps]
      summary: List apps with optional search and category filter
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: sort
          schema:
            type: string
            enum: [latest, trending, rating]
      responses:
        "200":
          description: List of apps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/App"
    post:
      operationId: createApp
      tags: [apps]
      summary: Create a new app (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/App"

  /apps/featured:
    get:
      operationId: listFeaturedApps
      tags: [apps]
      summary: Top featured apps
      responses:
        "200":
          description: Featured
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/App"

  /apps/trending:
    get:
      operationId: listTrendingApps
      tags: [apps]
      summary: Trending apps (most downloads in last days)
      responses:
        "200":
          description: Trending
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/App"

  /apps/latest:
    get:
      operationId: listLatestApps
      tags: [apps]
      summary: Latest added apps
      responses:
        "200":
          description: Latest
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/App"

  /apps/{id}:
    get:
      operationId: getApp
      tags: [apps]
      summary: Get an app by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: App
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/App"
    delete:
      operationId: deleteApp
      tags: [apps]
      summary: Delete an app (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

  /apps/{id}/download:
    post:
      operationId: registerDownload
      tags: [apps]
      summary: Register a download click and return the URL
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Download URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadResponse"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories with app counts
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /stats/overview:
    get:
      operationId: getStatsOverview
      tags: [stats]
      summary: Overview stats for admin/home
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsOverview"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status: { type: string }
      required: [status]

    App:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        description: { type: string }
        logoUrl: { type: string }
        screenshots:
          type: array
          items: { type: string }
        downloadUrl: { type: string }
        size: { type: string }
        version: { type: string }
        developer: { type: string }
        rating: { type: number }
        downloads: { type: integer }
        featured: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, name, category, description, logoUrl, screenshots, downloadUrl, size, version, developer, rating, downloads, featured, createdAt]

    CreateAppBody:
      type: object
      properties:
        name: { type: string }
        category: { type: string }
        description: { type: string }
        logoUrl: { type: string }
        screenshots:
          type: array
          items: { type: string }
        downloadUrl: { type: string }
        size: { type: string }
        version: { type: string }
        developer: { type: string }
        featured: { type: boolean }
      required: [name, category, description, logoUrl, screenshots, downloadUrl, size, version, developer]

    Category:
      type: object
      properties:
        slug: { type: string }
        name: { type: string }
        appCount: { type: integer }
      required: [slug, name, appCount]

    DownloadResponse:
      type: object
      properties:
        url: { type: string }
        downloads: { type: integer }
      required: [url, downloads]

    StatsOverview:
      type: object
      properties:
        totalApps: { type: integer }
        totalDownloads: { type: integer }
        totalCategories: { type: integer }
        averageRating: { type: number }
      required: [totalApps, totalDownloads, totalCategories, averageRating]