# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hitler Mod app store API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List apps with optional search and category filter
 */
export const ListAppsQueryParams = zod.object({
  search: zod.coerce.string().optional(),
  category: zod.coerce.string().optional(),
  sort: zod.enum(["latest", "trending", "rating"]).optional(),
});

export const ListAppsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  category: zod.string(),
  description: zod.string(),
  logoUrl: zod.string(),
  screenshots: zod.array(zod.string()),
  downloadUrl: zod.string(),
  size: zod.string(),
  version: zod.string(),
  developer: zod.string(),
  rating: zod.number(),
  downloads: zod.number(),
  featured: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const ListAppsResponse = zod.array(ListAppsResponseItem);

/**
 * @summary Create a new app (admin)
 */
export const CreateAppBody = zod.object({
  name: zod.string(),
  category: zod.string(),
  description: zod.string(),
  logoUrl: zod.string(),
  screenshots: zod.array(zod.string()),
  downloadUrl: zod.string(),
  size: zod.string(),
  version: zod.string(),
  developer: zod.string(),
  featured: zod.boolean().optional(),
});

/**
 * @summary Top featured apps
 */
export const ListFeaturedAppsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  category: zod.string(),
  description: zod.string(),
  logoUrl: zod.string(),
  screenshots: zod.array(zod.string()),
  downloadUrl: zod.string(),
  size: zod.string(),
  version: zod.string(),
  developer: zod.string(),
  rating: zod.number(),
  downloads: zod.number(),
  featured: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const ListFeaturedAppsResponse = zod.array(ListFeaturedAppsResponseItem);

/**
 * @summary Trending apps (most downloads in last days)
 */
export const ListTrendingAppsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  category: zod.string(),
  description: zod.string(),
  logoUrl: zod.string(),
  screenshots: zod.array(zod.string()),
  downloadUrl: zod.string(),
  size: zod.string(),
  version: zod.string(),
  developer: zod.string(),
  rating: zod.number(),
  downloads: zod.number(),
  featured: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const ListTrendingAppsResponse = zod.array(ListTrendingAppsResponseItem);

/**
 * @summary Latest added apps
 */
export const ListLatestAppsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  category: zod.string(),
  description: zod.string(),
  logoUrl: zod.string(),
  screenshots: zod.array(zod.string()),
  downloadUrl: zod.string(),
  size: zod.string(),
  version: zod.string(),
  developer: zod.string(),
  rating: zod.number(),
  downloads: zod.number(),
  featured: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const ListLatestAppsResponse = zod.array(ListLatestAppsResponseItem);

/**
 * @summary Get an app by id
 */
export const GetAppParams = zod.object({
  id: zod.coerce.string(),
});

export const GetAppResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  category: zod.string(),
  description: zod.string(),
  logoUrl: zod.string(),
  screenshots: zod.array(zod.string()),
  downloadUrl: zod.string(),
  size: zod.string(),
  version: zod.string(),
  developer: zod.string(),
  rating: zod.number(),
  downloads: zod.number(),
  featured: zod.boolean(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete an app (admin)
 */
export const DeleteAppParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Register a download click and return the URL
 */
export const RegisterDownloadParams = zod.object({
  id: zod.coerce.string(),
});

export const RegisterDownloadResponse = zod.object({
  url: zod.string(),
  downloads: zod.number(),
});

/**
 * @summary List categories with app counts
 */
export const ListCategoriesResponseItem = zod.object({
  slug: zod.string(),
  name: zod.string(),
  appCount: zod.number(),
});
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary Overview stats for admin/home
 */
export const GetStatsOverviewResponse = zod.object({
  totalApps: zod.number(),
  totalDownloads: zod.number(),
  totalCategories: zod.number(),
  averageRating: zod.number(),
});