# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hitler Mod app store API
 * OpenAPI spec version: 0.1.0
 */

export interface Category {
  slug: string;
  name: string;
  appCount: number;
}