# lib/api-zod/src/generated/types/createAppBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hitler Mod app store API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAppBody {
  name: string;
  category: string;
  description: string;
  logoUrl: string;
  screenshots: string[];
  downloadUrl: string;
  size: string;
  version: string;
  developer: string;
  featured?: boolean;
}