# lib/api-zod/src/generated/types/statsOverview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hitler Mod app store API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsOverview {
  totalApps: number;
  totalDownloads: number;
  totalCategories: number;
  averageRating: number;
}